requirements.txt
numpy>=1.24
pandas>=2.0
streamlit>=1.32
scikit-learn>=1.3
joblib>=1.3
requests>=2.31
fastapi>=0.111
uvicorn>=0.30
httpx>=0.28
xgboost>=3.2
lightgbm>=4.6
shap>=0.51
google-generativeai>=0.3.0
